requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
pydantic>=2.0
